Fix typedef size_t as unsigned int
(fatal side effect for Lua, #343)

+++ src/include/net/socket.h
@@ -19,33 +19,7 @@
 //#define AF_UNIX      1  /* Unix domain sockets */
 //#define AF_LOCAL     1  /* POSIX name for AF_UNIX */
 #define AF_INET      2  /* Internet IP Protocol */
-//#define AF_AX25      3  /* Amateur Radio AX.25 */
-//#define AF_IPX       4  /* Novell IPX */
-//#define AF_APPLETALK 5  /* AppleTalk DDP */
-//#define AF_NETROM    6  /* Amateur Radio NET/ROM */
-//#define AF_BRIDGE    7  /* Multiprotocol bridge */
-//#define AF_ATMPVC    8  /* ATM PVCs */
-//#define AF_X25       9  /* Reserved for X.25 project */
-//#define AF_INET6     10 /* IP version 6 */
-//#define AF_ROSE      11 /* Amateur Radio X.25 PLP */
-//#define AF_DECnet    12 /* Reserved for DECnet project */
-//#define AF_NETBEUI   13 /* Reserved for 802.2LLC project */
-//#define AF_SECURITY  14 /* Security callback pseudo AF */
-//#define AF_KEY       15 /* PF_KEY key management API */
-//#define AF_NETLINK   16
-//#define AF_ROUTE     16 /* Alias to emulate 4.4BSD (such as AF_NETLINK) */
 #define AF_PACKET    17 /* Packet family */
-//#define AF_ASH       18 /* Ash */
-//#define AF_ECONET    19 /* Acorn Econet */
-//#define AF_ATMSVC    20 /* ATM SVCs */
-//#define AF_SNA       22 /* Linux SNA Project (nutters!) */
-//#define AF_IRDA      23 /* IRDA sockets */
-//#define AF_PPPOX     24 /* PPPoX sockets */
-//#define AF_WANPIPE   25 /* Wanpipe API Sockets */
-//#define AF_LLC       26 /* Linux LLC */
-//#define AF_TIPC      30 /* TIPC sockets */
-//#define AF_BLUETOOTH 31 /* Bluetooth sockets */
-//#define AF_IUCV      32 /* IUCV sockets */
 #define AF_MAX       33 /* For now.. */
 
 /* Protocol families, same as address families. */
@@ -53,35 +27,8 @@
 //#define PF_UNIX		AF_UNIX
 //#define PF_LOCAL	AF_LOCAL
 #define PF_INET		AF_INET
-//#define PF_AX25		AF_AX25
-//#define PF_IPX		AF_IPX
-//#define PF_APPLETALK	AF_APPLETALK
-//#define	PF_NETROM	AF_NETROM
-//#define PF_BRIDGE	AF_BRIDGE
-//#define PF_ATMPVC	AF_ATMPVC
-//#define PF_X25		AF_X25
-//#define PF_INET6	AF_INET6
-//#define PF_ROSE		AF_ROSE
-//#define PF_DECnet	AF_DECnet
-//#define PF_NETBEUI	AF_NETBEUI
-//#define PF_SECURITY	AF_SECURITY
-//#define PF_KEY		AF_KEY
-//#define PF_NETLINK	AF_NETLINK
-//#define PF_ROUTE	AF_ROUTE
 #define PF_PACKET	AF_PACKET
-//#define PF_ASH		AF_ASH
-//#define PF_ECONET	AF_ECONET
-//#define PF_ATMSVC	AF_ATMSVC
-//#define PF_SNA		AF_SNA
-//#define PF_IRDA		AF_IRDA
-//#define PF_PPPOX	AF_PPPOX
-//#define PF_WANPIPE	AF_WANPIPE
-//#define PF_LLC		AF_LLC
-//#define PF_TIPC		AF_TIPC
-//#define PF_BLUETOOTH	AF_BLUETOOTH
-//#define PF_IUCV		AF_IUCV
 #define PF_MAX		AF_MAX
-
 
 #define PNET_GRAPH   40
 
@@ -212,7 +159,7 @@ extern int bind(int sockfd, const struct sockaddr *addr, socklen_t addrlen);
  * @param daddrlen size of daddr
  * @return the number of characters sent. On error, result < 0.
  */
-extern ssize_t sendto(int sockfd, const void *buf, int len, int flags,
+extern ssize_t sendto(int sockfd, const void *buf, size_t len, int flags,
 		const struct sockaddr *daddr, socklen_t daddrlen);
 
 /**
@@ -225,7 +172,7 @@ extern ssize_t sendto(int sockfd, const void *buf, int len, int flags,
  * @param daddrlen size of daddr
  * @return the number of bytes received, or result < 0 if an error occurred.
  */
-extern ssize_t recvfrom(int sockfd, void *buf, int len, int flags,
+extern ssize_t recvfrom(int sockfd, void *buf, size_t len, int flags,
 		struct sockaddr *daddr, socklen_t *daddrlen);
 
 #if 0
